Replace `ShamatarLib` with `Nethermind.Crypto.Pairings` package
