+++ src/Nethermind/Directory.Packages.props
@@ -40,6 +40,7 @@
     <PackageVersion Include="Microsoft.Extensions.ObjectPool" Version="7.0.3" />
     <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.4.1" />
     <PackageVersion Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.17.0" />
+    <PackageVersion Include="Nethermind.Crypto.Pairings" Version="1.0.0-preview.2" />
     <PackageVersion Include="Nethermind.Crypto.SecP256k1" Version="1.0.0" />
     <PackageVersion Include="Nethermind.DotNetty.Buffers" Version="1.0.0" />
     <PackageVersion Include="Nethermind.DotNetty.Handlers" Version="1.0.0" />
