+++ src/Nethermind/Nethermind.Evm.Test/BlsMultiMulG2PrecompileTests.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using FluentAssertions;
 using Nethermind.Core.Extensions;
 using Nethermind.Evm.Precompiles;
-using Nethermind.Evm.Precompiles.Bls.Shamatar;
+using Nethermind.Evm.Precompiles.Bls;
 using NUnit.Framework;
 using static Nethermind.Specs.Forks.MuirGlacier;
