+++ src/Nethermind/Nethermind.Evm/Nethermind.Evm.csproj
@@ -14,6 +14,7 @@
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="FastEnum" />
+    <PackageReference Include="Nethermind.Crypto.Pairings" />
     <PackageReference Include="Nethermind.Gmp" />
   </ItemGroup>
 </Project>
