+++ src/Nethermind/Nethermind.Evm/Precompiles/Bls/MapToG1Precompile.cs
@@ -4,26 +4,26 @@
 using System;
 using Nethermind.Core;
 using Nethermind.Core.Specs;
-using Nethermind.Crypto.Bls;
+using Nethermind.Crypto;
 
-namespace Nethermind.Evm.Precompiles.Bls.Shamatar
+namespace Nethermind.Evm.Precompiles.Bls
 {
     /// <summary>
     /// https://eips.ethereum.org/EIPS/eip-2537
     /// </summary>
-    public class G1MulPrecompile : IPrecompile
+    public class MapToG1Precompile : IPrecompile
     {
-        public static IPrecompile Instance = new G1MulPrecompile();
+        public static IPrecompile Instance = new MapToG1Precompile();
 
-        private G1MulPrecompile()
+        private MapToG1Precompile()
         {
         }
 
-        public Address Address { get; } = Address.FromNumber(11);
+        public Address Address { get; } = Address.FromNumber(17);
 
         public long BaseGasCost(IReleaseSpec releaseSpec)
         {
-            return 12000L;
+            return 5500L;
         }
 
         public long DataGasCost(in ReadOnlyMemory<byte> inputData, IReleaseSpec releaseSpec)
@@ -33,7 +33,7 @@ namespace Nethermind.Evm.Precompiles.Bls.Shamatar
 
         public (ReadOnlyMemory<byte>, bool) Run(in ReadOnlyMemory<byte> inputData, IReleaseSpec releaseSpec)
         {
-            const int expectedInputLength = 2 * BlsParams.LenFp + BlsParams.LenFr;
+            const int expectedInputLength = 64;
             if (inputData.Length != expectedInputLength)
             {
                 return (Array.Empty<byte>(), false);
@@ -44,8 +44,8 @@ namespace Nethermind.Evm.Precompiles.Bls.Shamatar
 
             (byte[], bool) result;
 
-            Span<byte> output = stackalloc byte[2 * BlsParams.LenFp];
-            bool success = ShamatarLib.BlsG1Mul(inputData.Span, output);
+            Span<byte> output = stackalloc byte[128];
+            bool success = Pairings.BlsMapToG1(inputData.Span, output);
             if (success)
             {
                 result = (output.ToArray(), true);
