+++ src/Nethermind/Nethermind.Evm/Precompiles/Snarks/Bn254AddPrecompile.cs
@@ -4,22 +4,22 @@
 using System;
 using Nethermind.Core;
 using Nethermind.Core.Specs;
-using Nethermind.Crypto.Bls;
+using Nethermind.Crypto;
 
-namespace Nethermind.Evm.Precompiles.Snarks.Shamatar
+namespace Nethermind.Evm.Precompiles.Snarks
 {
     /// <summary>
-    /// https://github.com/herumi/mcl/blob/master/api.md
+    /// https://github.com/matter-labs/eip1962/blob/master/eip196_header.h
     /// </summary>
-    public class Bn256MulPrecompile : IPrecompile
+    public class Bn254AddPrecompile : IPrecompile
     {
-        public static IPrecompile Instance = new Bn256MulPrecompile();
+        public static IPrecompile Instance = new Bn254AddPrecompile();
 
-        public Address Address { get; } = Address.FromNumber(7);
+        public Address Address { get; } = Address.FromNumber(6);
 
         public long BaseGasCost(IReleaseSpec releaseSpec)
         {
-            return releaseSpec.IsEip1108Enabled ? 6000L : 40000L;
+            return releaseSpec.IsEip1108Enabled ? 150L : 500L;
         }
 
         public long DataGasCost(in ReadOnlyMemory<byte> inputData, IReleaseSpec releaseSpec)
@@ -27,14 +27,14 @@ namespace Nethermind.Evm.Precompiles.Snarks.Shamatar
             return 0L;
         }
 
-        public (ReadOnlyMemory<byte>, bool) Run(in ReadOnlyMemory<byte> inputData, IReleaseSpec releaseSpec)
+        public unsafe (ReadOnlyMemory<byte>, bool) Run(in ReadOnlyMemory<byte> inputData, IReleaseSpec releaseSpec)
         {
-            Metrics.Bn256MulPrecompile++;
-            Span<byte> inputDataSpan = stackalloc byte[96];
+            Metrics.Bn254AddPrecompile++;
+            Span<byte> inputDataSpan = stackalloc byte[128];
             inputData.PrepareEthInput(inputDataSpan);
 
             Span<byte> output = stackalloc byte[64];
-            bool success = ShamatarLib.Bn256Mul(inputDataSpan, output);
+            bool success = Pairings.Bn254Add(inputDataSpan, output);
 
             (byte[], bool) result;
             if (success)
