+++ src/Nethermind/Nethermind.Precompiles.Benchmark/Bn254PairingBenchmark.cs
@@ -2,19 +2,18 @@
 // SPDX-License-Identifier: LGPL-3.0-only
 
 using System.Collections.Generic;
-using BenchmarkDotNet.Attributes;
-using BenchmarkDotNet.Jobs;
 using Nethermind.Evm.Precompiles;
+using Nethermind.Evm.Precompiles.Snarks;
 
 namespace Nethermind.Precompiles.Benchmark
 {
-    public class Bn256MulBenchmark : PrecompileBenchmarkBase
+    public class Bn254PairingBenchmark : PrecompileBenchmarkBase
     {
         protected override IEnumerable<IPrecompile> Precompiles => new[]
         {
-            Evm.Precompiles.Snarks.Shamatar.Bn256MulPrecompile.Instance
+            Bn254PairingPrecompile.Instance
         };
 
-        protected override string InputsDirectory => "bnmul";
+        protected override string InputsDirectory => "bnpair";
     }
 }
