+++ src/Nethermind/nuget.config
@@ -3,14 +3,14 @@
   <packageSources>
     <clear />
     <add key="nuget.org" value="https://api.nuget.org/v3/index.json" />
-    <!--<add key="nugettest.org" value="https://apiint.nugettest.org/v3/index.json" />-->
+    <add key="nugettest.org" value="https://apiint.nugettest.org/v3/index.json" />
   </packageSources>
-  <!--<packageSourceMapping>
+  <packageSourceMapping>
     <packageSource key="nuget.org">
       <package pattern="*" />
     </packageSource>
     <packageSource key="nugettest.org">
-      <package pattern="" />
+      <package pattern="Nethermind.Crypto.Pairings" />
     </packageSource>
-  </packageSourceMapping>-->
+  </packageSourceMapping>
 </configuration>
